+++ wemake_python_styleguide/visitors/ast/annotations.py
@@ -1,0 +1,58 @@
+# -*- coding: utf-8 -*-
+
+import ast
+
+from wemake_python_styleguide.types import AnyFunctionDef, final
+from wemake_python_styleguide.violations.consistency import (
+    MultilineFunctionAnnotationViolation,
+)
+from wemake_python_styleguide.visitors.base import BaseNodeVisitor
+from wemake_python_styleguide.visitors.decorators import alias
+
+
+@final
+@alias('visit_any_function', (
+    'visit_FunctionDef',
+    'visit_AsyncFunctionDef',
+))
+class WrongAnnotationVisitor(BaseNodeVisitor):
+    """Ensures that annotations are used correctly."""
+
+    def _check_arg_annotation(self, node: ast.arg) -> None:
+        for sub_node in ast.walk(node):
+            lineno = getattr(sub_node, 'lineno', None)
+            if lineno and lineno != node.lineno:
+                self.add_violation(MultilineFunctionAnnotationViolation(node))
+                return
+
+    def _check_return_annotation(self, node: ast.FunctionDef) -> None:
+        if not node.returns:
+            return
+
+        for sub_node in ast.walk(node.returns):
+            lineno = getattr(sub_node, 'lineno', None)
+            if lineno and lineno != node.returns.lineno:
+                self.add_violation(MultilineFunctionAnnotationViolation(node))
+                return
+
+    def visit_any_function(self, node: AnyFunctionDef) -> None:
+        """
+        Checks return type annotations.
+
+        Raises:
+            MultilineFunctionAnnotationViolation
+
+        """
+        self._check_return_annotation(node)
+        self.generic_visit(node)
+
+    def visit_arg(self, node: ast.arg) -> None:
+        """
+        Checks arguments annotations.
+
+        Raises:
+            MultilineFunctionAnnotationViolation
+
+        """
+        self._check_arg_annotation(node)
+        self.generic_visit(node)